menambahkan class baru: PersonalInfo
telalu banyak variable yang tidak dibuat setter getternya(tidak dipanggil dalam fungsi apapun) dan memiliki hubungan yang erat
